+++ src/replicon_core/replication_rules.rs
@@ -190,7 +190,7 @@ pub(crate) struct ReplicationInfo {
 }
 
 /// Marks entity for replication.
-#[derive(Component, Clone, Copy, Default)]
+#[derive(Component, Clone, Copy, Default, Reflect)]
 pub struct Replication;
 
 /// Replication will be ignored for `T` if this component is present on the same entity.